Make into interfaces to remove lwjgl dependencies from GUIManager and UIDisplayRenderer
Provide headless versions
